data-sources/requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
kafka-python==2.0.1
loguru==0.5.1
pathspec==0.8.0
psycopg2==2.8.5
regex==2020.6.8
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.5.4
requests~=2.24.0
cryptography~=2.9.2
httpretty~=1.1.3
ripemd-hash==1.0.0